requirements.txt
fastapi
uvicorn[standard]
python-docx
multipart
python-dotenv
pydantic-settings
pywin32
supabase
gunicorn
requests
